+++ psyneulink/core/components/functions/__init__.py
@@ -1,8 +1,14 @@
 from . import function
 from . import userdefinedfunction
+from . import combinationfunctions
+from . import interfacefunctions
 
 from .function import *
 from .userdefinedfunction import *
+from .combinationfunctions import *
+from .interfacefunctions import *
 
 __all__ = list(function.__all__)
 __all__.extend(userdefinedfunction.__all__)
+__all__.extend(combinationfunctions.__all__)
+__all__.extend(interfacefunctions.__all__)
