+++ psyneulink/core/components/functions/function.py
@@ -19,6 +19,10 @@ Combination Functions:
   * `CombineMeans`
   * `PredictionErrorDeltaFunction`
 
+Interface Functions:
+  * Identity
+  * InterfaceStateMap
+
 Transfer Functions:
   * `Linear`
   * `Exponential`
@@ -218,9 +222,10 @@ from psyneulink.core.globals.context import ContextFlags
 from psyneulink.core.globals.defaults import MPI_IMPLEMENTATION
 from psyneulink.core.globals.keywords import ACCUMULATOR_INTEGRATOR_FUNCTION, ADAPTIVE_INTEGRATOR_FUNCTION, ALL, ARGUMENT_THERAPY_FUNCTION, AUTO_ASSIGN_MATRIX, BACKPROPAGATION_FUNCTION, BETA, BIAS, BUFFER_FUNCTION, \
     CONSTANT_INTEGRATOR_FUNCTION, CONTRASTIVE_HEBBIAN_FUNCTION, CORRELATION, CROSS_ENTROPY, DECAY, DEFAULT_VARIABLE, DIFFERENCE, DISTANCE_FUNCTION, DISTANCE_METRICS, DIST_FUNCTION_TYPE, DIST_MEAN, DIST_SHAPE, DRIFT_DIFFUSION_INTEGRATOR_FUNCTION, DistanceMetrics, ENERGY, ENTROPY, EUCLIDEAN, EXAMPLE_FUNCTION_TYPE, EXPONENTIAL, EXPONENTIAL_DIST_FUNCTION, EXPONENTIAL_FUNCTION, \
-    FHN_INTEGRATOR_FUNCTION, FULL_CONNECTIVITY_MATRIX, FUNCTION, FUNCTION_OUTPUT_TYPE, FUNCTION_OUTPUT_TYPE_CONVERSION, GAIN, GAMMA_DIST_FUNCTION, GAUSSIAN, GAUSSIAN_FUNCTION, GRADIENT_OPTIMIZATION_FUNCTION, GRID_SEARCH_FUNCTION, HAS_INITIALIZERS, HEBBIAN_FUNCTION, HIGH, HOLLOW_MATRIX, IDENTITY_FUNCTION, IDENTITY_MATRIX, INCREMENT, INITIALIZER, INPUT_STATES, INTEGRATOR_FUNCTION, INTEGRATOR_FUNCTION_TYPE, INTERCEPT, KOHONEN_FUNCTION, LCAMechanism_INTEGRATOR_FUNCTION, LEAK, LEARNING_FUNCTION_TYPE, LEARNING_RATE, LINEAR, \
+    FHN_INTEGRATOR_FUNCTION, FULL_CONNECTIVITY_MATRIX, FUNCTION, FUNCTION_OUTPUT_TYPE, FUNCTION_OUTPUT_TYPE_CONVERSION, GAIN, GAMMA_DIST_FUNCTION, GAUSSIAN, GAUSSIAN_FUNCTION, GRADIENT_OPTIMIZATION_FUNCTION, GRID_SEARCH_FUNCTION, HAS_INITIALIZERS, HEBBIAN_FUNCTION, HIGH, HOLLOW_MATRIX, \
+    IDENTITY_MATRIX, INCREMENT, INITIALIZER, INPUT_STATES, INTEGRATOR_FUNCTION, INTEGRATOR_FUNCTION_TYPE, INTERCEPT, KOHONEN_FUNCTION, LCAMechanism_INTEGRATOR_FUNCTION, LEAK, LEARNING_FUNCTION_TYPE, LEARNING_RATE, LINEAR, \
     LINEAR_FUNCTION, LINEAR_MATRIX_FUNCTION, LOGISTIC_FUNCTION, LOW, MATRIX, MATRIX_KEYWORD_NAMES, MATRIX_KEYWORD_VALUES, MAX_ABS_DIFF, MAX_ABS_INDICATOR, MAX_ABS_VAL, MAX_INDICATOR, MAX_VAL, NAME, NOISE, NORMAL_DIST_FUNCTION, OBJECTIVE_FUNCTION_TYPE, OFFSET, ONE_HOT_FUNCTION, OPERATION, OPTIMIZATION_FUNCTION_TYPE, ORNSTEIN_UHLENBECK_INTEGRATOR_FUNCTION, OUTPUT_STATES, OUTPUT_TYPE, OWNER, PARAMETER_STATE_PARAMS, \
-    PER_ITEM, PROB, PROB_INDICATOR, RANDOM_CONNECTIVITY_MATRIX, RATE, RECEIVER, RELU_FUNCTION, RL_FUNCTION, SCALE, SELECTION_FUNCTION_TYPE, SIMPLE_INTEGRATOR_FUNCTION, SLOPE, SOFTMAX_FUNCTION, STABILITY_FUNCTION, STANDARD_DEVIATION, STATE_MAP_FUNCTION, \
+    PER_ITEM, PROB, PROB_INDICATOR, RANDOM_CONNECTIVITY_MATRIX, RATE, RECEIVER, RELU_FUNCTION, RL_FUNCTION, SCALE, SELECTION_FUNCTION_TYPE, SIMPLE_INTEGRATOR_FUNCTION, SLOPE, SOFTMAX_FUNCTION, STABILITY_FUNCTION, STANDARD_DEVIATION, \
     TDLEARNING_FUNCTION, TIME_STEP_SIZE, TRANSFER_FUNCTION_TYPE, UNIFORM_DIST_FUNCTION, UTILITY_INTEGRATOR_FUNCTION, VALUE, VARIABLE, VARIANCE, WALD_DIST_FUNCTION, \
     X_0, kwComponentCategory, kwPreferenceSetName
 from psyneulink.core.globals.parameters import Param
@@ -1171,328 +1176,6 @@ class ArgumentTherapy(Function_Base):
 # region ****************************************   FUNCTIONS   ********************************************************
 # endregion
 
-
-# region ***********************************  INTERFACE FUNCTIONS ***********************************************
-
-class InterfaceFunction(Function_Base):
-    """Simple functions for CompositionInterfaceMechanisms
-    """
-    componentType = TRANSFER_FUNCTION_TYPE
-
-    def __init__(self,
-                 default_variable,
-                 params,
-                 owner,
-                 prefs,
-                 context):
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=context)
-
-
-class Identity(InterfaceFunction):  # -------------------------------------------------------------------------------------
-    """
-    Identity(                \
-             default_variable, \
-             params=None,      \
-             owner=None,       \
-             name=None,        \
-             prefs=None        \
-            )
-
-    .. _Identity:
-
-    Returns variable
-
-    Arguments
-    ---------
-
-    variable : number or np.array : default ClassDefaults.variable
-        specifies a template for the value to be transformed.
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    Attributes
-    ----------
-
-    variable : number or np.array
-        contains value to be transformed.
-
-    owner : Component
-        `component <Component>` to which the Function has been assigned.
-
-    name : str
-        the name of the Function; if it is not specified in the **name** argument of the constructor, a
-        default is assigned by FunctionRegistry (see `Naming` for conventions used for default and duplicate names).
-
-    prefs : PreferenceSet or specification dict : Function.classPreferences
-        the `PreferenceSet` for function; if it is not specified in the **prefs** argument of the Function's
-        constructor, a default is assigned using `classPreferences` defined in __init__.py (see :doc:`PreferenceSet
-        <LINK>` for details).
-    """
-
-    componentName = IDENTITY_FUNCTION
-
-    classPreferences = {
-        kwPreferenceSetName: 'LinearClassPreferences',
-        kpReportOutputPref: PreferenceEntry(False, PreferenceLevel.INSTANCE),
-    }
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-    # paramClassDefaults.update({
-    #     FUNCTION_OUTPUT_TYPE_CONVERSION: False,
-    #     PARAMETER_STATE_PARAMS: None
-    # })
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-        # self.functionOutputType = None
-
-    def function(
-        self,
-        variable=None,
-        execution_id=None,
-        params=None,
-        context=None
-    ):
-        """
-        Return: `variable <Identity.variable>`
-
-        Arguments
-        ---------
-
-        variable : number or np.array : default ClassDefaults.variable
-           a single value or array to be transformed.
-
-        params : Dict[param keyword: param value] : default None
-            a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-            function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-            arguments of the constructor.
-
-
-        Returns
-        -------
-
-        variable : number or np.array
-
-        """
-
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-        # outputType = self.functionOutputType
-
-        return variable
-
-    def _get_input_struct_type(self,ctx):
-        #FIXME: Workaround for CompositionInterfaceMechanism that
-        #       does not udpate its instance_defaults shape
-        from psyneulink.core.components.mechanisms.processing.compositioninterfacemechanism import CompositionInterfaceMechanism
-        if isinstance(self.owner, CompositionInterfaceMechanism):
-            variable = [state.instance_defaults.value for state in self.owner.input_states]
-            # Python list does not care about ndarrays of different lengths
-            # we do care, so convert to tuple to create struct
-            if all(type(x) == np.ndarray for x in variable) and not all(len(x) == len(variable[0]) for x in variable):
-                variable = tuple(variable)
-
-            return ctx.convert_python_struct_to_llvm_ir(variable)
-        return ctx.get_input_struct_type(super())
-
-    def _get_output_struct_type(self, ctx):
-        #FIXME: Workaround for CompositionInterfaceMechanism that
-        #       does not udpate its instance_defaults shape
-        from psyneulink.core.components.mechanisms.processing.compositioninterfacemechanism import CompositionInterfaceMechanism
-        if isinstance(self.owner, CompositionInterfaceMechanism):
-            return ctx.get_input_struct_type(self)
-        return ctx.get_output_struct_type(super())
-
-    def _gen_llvm_function_body(self, ctx, builder, _1, _2, arg_in, arg_out):
-        val = builder.load(arg_in)
-        builder.store(val, arg_out)
-        return builder
-
-
-class InterfaceStateMap(InterfaceFunction):
-    """
-    Identity(                \
-             default_variable, \
-             params=None,      \
-             owner=None,       \
-             name=None,        \
-             prefs=None        \
-            )
-
-    .. _Identity:
-
-    Returns `variable <InterfaceStateMap.variable>`.
-
-    Arguments
-    ---------
-
-    default_variable : number or np.array : default ClassDefaults.variable
-        specifies a template for the value to be transformed.
-
-    params : Dict[param keyword: param value] : default None
-        a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-        function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-        arguments of the constructor.
-
-    owner : Component
-        `component <Component>` to which to assign the Function.
-
-    name : str : default see `name <Function.name>`
-        specifies the name of the Function.
-
-    prefs : PreferenceSet or specification dict : default Function.classPreferences
-        specifies the `PreferenceSet` for the Function (see `prefs <Function_Base.prefs>` for details).
-
-    Attributes
-    ----------
-
-    variable : number or np.array
-        contains value to be transformed.
-
-    owner : Component
-        `component <Component>` to which the Function has been assigned.
-
-    name : str
-        the name of the Function; if it is not specified in the **name** argument of the constructor, a
-        default is assigned by FunctionRegistry (see `Naming` for conventions used for default and duplicate names).
-
-    prefs : PreferenceSet or specification dict : Function.classPreferences
-        the `PreferenceSet` for function; if it is not specified in the **prefs** argument of the Function's
-        constructor, a default is assigned using `classPreferences` defined in __init__.py (see :doc:`PreferenceSet
-        <LINK>` for details).
-    """
-
-    componentName = STATE_MAP_FUNCTION
-
-    classPreferences = {
-        kwPreferenceSetName: 'LinearClassPreferences',
-        kpReportOutputPref: PreferenceEntry(False, PreferenceLevel.INSTANCE),
-    }
-
-    paramClassDefaults = Function_Base.paramClassDefaults.copy()
-    paramClassDefaults.update({
-        FUNCTION_OUTPUT_TYPE_CONVERSION: True,
-        PARAMETER_STATE_PARAMS: None
-    })
-
-    @tc.typecheck
-    def __init__(self,
-                 default_variable=None,
-                 corresponding_input_state=None,
-                 params=None,
-                 owner=None,
-                 prefs: is_pref_set = None):
-
-        # Assign args to params and functionParams dicts (kwConstants must == arg names)
-        params = self._assign_args_to_param_dicts(corresponding_input_state=corresponding_input_state,
-                                                  params=params)
-
-        super().__init__(default_variable=default_variable,
-                         params=params,
-                         owner=owner,
-                         prefs=prefs,
-                         context=ContextFlags.CONSTRUCTOR)
-
-        # self.functionOutputType = None
-
-    def function(
-        self,
-        variable=None,
-        execution_id=None,
-        params=None,
-        context=None
-    ):
-        """
-        Return: The item of `value <InterfaceStateMap.value>` whose index corresponds to the index of
-        `corresponding_input_state <InterfaceStateMap.corresponding_input_state>` in `input_states
-        <InterfaceStateMap.input_states>`
-
-        Arguments
-        ---------
-
-        variable : number or np.array : default ClassDefaults.variable
-           a single value or array to be transformed.
-
-        corresponding_input_state : InputState : default None
-            the InputState on the owner CompositionInterfaceMechanism to which this OutputState corresponds
-
-        params : Dict[param keyword: param value] : default None
-            a `parameter dictionary <ParameterState_Specification>` that specifies the parameters for the
-            function.  Values specified for parameters in the dictionary override any assigned to those parameters in
-            arguments of the constructor.
-
-
-        Returns
-        -------
-
-        The item of `value <InterfaceStateMap.value>` whose index corresponds to the index of
-        `corresponding_input_state <InterfaceStateMap.corresponding_input_state>` in `input_states
-        <InterfaceStateMap.input_states>`
-
-        """
-        variable = self._check_args(variable=variable, execution_id=execution_id, params=params, context=context)
-
-        index = self.corresponding_input_state.position_in_mechanism
-
-        if self.corresponding_input_state.owner.parameters.value.get(execution_id) is not None:
-
-            # If CIM's variable does not match its value, then a new pair of states was added since the last execution
-            if not np.shape(self.corresponding_input_state.owner.get_input_values(execution_id)) == np.shape(self.corresponding_input_state.owner.parameters.value.get(execution_id)):
-                return self.corresponding_input_state.owner.instance_defaults.variable[index]
-
-            # If the variable is 1D (e.g. [0. , 0.], NOT [[0. , 0.]]), and the index is 0, then return whole variable
-            # np.atleast_2d fails in cases like var = [[0., 0.], [0.]] (transforms it to [[[0., 0.], [0.]]])
-            if index == 0:
-                if not isinstance(variable[0], (list, np.ndarray)):
-                    return variable
-            return variable[index]
-        # CIM value = None, use CIM's default variable instead
-        return self.corresponding_input_state.owner.instance_defaults.variable[index]
-
-    def _get_input_struct_type(self, ctx):
-        #FIXME: Workaround for CompositionInterfaceMechanism that
-        #       does not update its instance_defaults shape
-        from psyneulink.core.components.mechanisms.processing.compositioninterfacemechanism import CompositionInterfaceMechanism
-        if hasattr(self.owner, 'owner') and isinstance(self.owner.owner, CompositionInterfaceMechanism):
-            return ctx.get_output_struct_type(self.owner.owner.function_object)
-        return ctx.get_input_struct_type(super())
-
-    def _gen_llvm_function_body(self, ctx, builder, _1, _2, arg_in, arg_out):
-        index = self.corresponding_input_state.position_in_mechanism
-        val = builder.load(builder.gep(arg_in, [ctx.int32_ty(0), ctx.int32_ty(index)]))
-        builder.store(val, arg_out)
-        return builder
-
-
-# endregion
 
 # region ***********************************  TRANSFER FUNCTIONS  ***********************************************
