+++ psyneulink/core/compositions/composition.py
@@ -58,7 +58,7 @@ from llvmlite import ir
 from psyneulink.core import llvm as pnlvm
 
 from psyneulink.core.components.component import ComponentsMeta, function_type
-from psyneulink.core.components.functions.function import InterfaceStateMap
+from psyneulink.core.components.functions.interfacefunctions import InterfaceStateMap
 from psyneulink.core.components.mechanisms.processing.compositioninterfacemechanism import CompositionInterfaceMechanism
 from psyneulink.core.components.mechanisms.processing.objectivemechanism import ObjectiveMechanism
 from psyneulink.core.components.projections.modulatory.modulatoryprojection import ModulatoryProjection_Base
