+++ tests/composition/test_interfaces.py
@@ -6,7 +6,8 @@ from timeit import timeit
 import numpy as np
 import pytest
 
-from psyneulink.core.components.functions.function import Identity, Linear, SimpleIntegrator
+from psyneulink.core.components.functions.function import Linear, SimpleIntegrator
+from psyneulink.core.components.functions.interfacefunctions import Identity
 from psyneulink.core.components.mechanisms.processing.compositioninterfacemechanism import CompositionInterfaceMechanism
 from psyneulink.core.components.mechanisms.processing.integratormechanism import IntegratorMechanism
 from psyneulink.core.components.mechanisms.processing.processingmechanism import ProcessingMechanism
