+++ tests/functions/test_identify.py
@@ -1,5 +1,6 @@
 import numpy as np
 import psyneulink.core.components.functions.function as Function
+import psyneulink.core.components.functions.interfacefunctions
 import psyneulink.core.globals.keywords as kw
 import pytest
 
@@ -9,7 +10,7 @@ import pytest
 @pytest.mark.benchmark
 def test_basic(size, benchmark):
     variable = np.random.rand(size)
-    f = Function.Identity(default_variable=variable)
+    f = psyneulink.core.components.functions.interfacefunctions.Identity(default_variable=variable)
     res = benchmark(f.function, variable)
     assert np.allclose(res, variable)
 
@@ -20,6 +21,6 @@ def test_basic(size, benchmark):
 @pytest.mark.benchmark
 def test_llvm(size, benchmark):
     variable = np.random.rand(size)
-    f = Function.Identity(default_variable=variable)
+    f = psyneulink.core.components.functions.interfacefunctions.Identity(default_variable=variable)
     res = benchmark(f.bin_function, variable)
     assert np.allclose(res, variable)
